Made the port configurable.
To do this the basic following changes where needed:
the nexus-test-harness-config was change from only containing files that would go in the runtime/apps/nexus/config directory to being an overlay of nexus.  This way any file (we should only override config files) can be changed.

The default port of the web app was changed to 8087 (at the off chance we collide with an already running instance of Nexus)

The unit tests where updated to get the nexus base URL from a props file (filtered at build time)

git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk/nexus/nexus-test-harness/nexus-test-harness-launcher@354 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab